5381: create list details files
