+++ .docker/wiremock/publizon/mappings/v1_user_reservations-1b8c0935-12c7-43c9-9a49-3e591cc079bb.json
@@ -1,22 +1,22 @@
 {
-  "id": "1b8c0935-12c7-43c9-9a49-3e591cc079bb",
-  "name": "Reservations (list)",
-  "request": {
-    "url": "/v1/user/reservations",
-    "method": "GET"
+  "id" : "1b8c0935-12c7-43c9-9a49-3e591cc079bb",
+  "name" : "Reservations (list)",
+  "request" : {
+    "url" : "/v1/user/reservations",
+    "method" : "GET"
   },
-  "response": {
-    "status": 200,
-    "body": "{\n  \"reservations\": [\n    {\n      \"productId\": \"0ddd10d0-d69f-4734-8a27-ac4546f4b912\",\n      \"identifier\": \"9788711321683\",\n      \"createdDateUtc\": \"2022-08-16T10:52:39.932Z\",\n      \"status\": 1,\n      \"productTitle\": \"Bargums synder\",\n      \"expireDateUtc\": null,\n      \"expectedRedeemDateUtc\": \"2023-01-27T19:37:15.63Z\"\n    },\n    {\n      \"productId\": \"0ddd10d0-d69f-4734-8a27-ac4546f4b912\",\n      \"identifier\": \"9788740047905\",\n      \"createdDateUtc\": \"2022-08-16T10:52:39.932Z\",\n      \"status\": 2,\n      \"productTitle\": \"Bargums synder\",\n      \"expireDateUtc\": null,\n      \"expectedRedeemDateUtc\": \"2023-01-27T19:37:15.63Z\"\n    }\n  ],\n  \"code\": 101,\n  \"message\": \"OK\"\n}",
-    "headers": {
-      "content-type": "application/json; charset=utf-8",
-      "date": "{{now format='yyyy-MM-dd HH:mm:ssZ'}}",
-      "strict-transport-security": "max-age=15552000"
+  "response" : {
+    "status" : 200,
+    "body" : "{\n  \"reservations\": [\n    {\n      \"productId\": \"0ddd10d0-d69f-4734-8a27-ac4546f4b912\",\n      \"identifier\": \"9788711321683\",\n      \"createdDateUtc\": \"2022-08-16T10:52:39.932Z\",\n      \"status\": 1,\n      \"productTitle\": \"Bargums synder\",\n      \"expireDateUtc\": \"2023-01-27T19:37:15.63Z\",\n      \"expectedRedeemDateUtc\": \"2023-01-27T19:37:15.63Z\"\n    },\n    {\n      \"productId\": \"0ddd10d0-d69f-4734-8a27-ac4546f4b912\",\n      \"identifier\": \"9788740047905\",\n      \"createdDateUtc\": \"2022-08-16T10:52:39.932Z\",\n      \"status\": 2,\n      \"productTitle\": \"Bargums synder\",\n      \"expireDateUtc\": \"2023-01-27T19:37:15.63Z\",\n      \"expectedRedeemDateUtc\": \"2023-01-27T19:37:15.63Z\"\n    }\n  ],\n  \"code\": 101,\n  \"message\": \"OK\"\n}",
+    "headers" : {
+      "content-type" : "application/json; charset=utf-8",
+      "date" : "{{now format='yyyy-MM-dd HH:mm:ssZ'}}",
+      "strict-transport-security" : "max-age=15552000"
     }
   },
-  "uuid": "1b8c0935-12c7-43c9-9a49-3e591cc079bb",
-  "persistent": true,
-  "priority": 5,
-  "insertionIndex": 23,
-  "postServeActions": []
+  "uuid" : "1b8c0935-12c7-43c9-9a49-3e591cc079bb",
+  "persistent" : true,
+  "priority" : 5,
+  "insertionIndex" : 23,
+  "postServeActions" : [ ]
 }
